move shared helper functions from load & delete into utils
